+++ web/apps/photos/src/utils/magicMetadata/index.ts
@@ -1,8 +1,7 @@
 import { sharedCryptoWorker } from "@/base/crypto";
 import type { Collection } from "@/media/collection";
 import { ItemVisibility } from "@/media/file-metadata";
-import { EnteFile } from "@/new/photos/types/file";
-import { MagicMetadataCore } from "@/new/photos/types/magicMetadata";
+import { EnteFile, MagicMetadataCore } from "@/new/photos/types/file";
 
 export function isArchivedFile(item: EnteFile): boolean {
     if (!item || !item.magicMetadata || !item.magicMetadata.data) {
