--- web/packages/new/photos/types/magicMetadata.ts
@@ -1,8 +1,0 @@
-export interface MagicMetadataCore<T> {
-    version: number;
-    count: number;
-    header: string;
-    data: T;
-}
-
-export type EncryptedMagicMetadata = MagicMetadataCore<string>;
